requirements.txt
streamlit>=1.40,<2.0
groq>=0.37,<1.0
pydantic>=2.9,<3.0
python-dotenv>=1.0,<2.0
networkx>=3.3,<4.0
numpy>=2.0,<3.0
pandas>=2.2,<3.0
plotly>=5.24,<6.0
